requirements.txt
python-dotenv==1.1.1
python-dateutil==2.8.2
pydantic>=2,<3
pydantic-settings>=2,<3
pyyaml==6.0.2
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.9
jinja2==3.1.6
openai==1.30.1
httpx<0.27
pytest==8.4.1
black==25.1.0
ics==0.7.2
google-api-python-client==2.177.0
google-auth==2.40.3
